(finally) Fixed InstanceCount issue(s)
